тест: арифметические функции

+++ Vvedenie/Task_1_2_7.java
@@ -3,7 +3,7 @@ package Vvedenie;
 public class Task_1_2_7 {
     public double priceCalculation(double price, int count) {
 
-        int sumPrice = (int) (price * count);
+        double sumPrice = price * count;
 
         System.out.println("Сумма стоимости товаров = " + sumPrice);
